Fix issue #327
Comments are now extracted as "begin" and "end" comment tokens by the lexer. The content
of the comment is extracted as a text token.
A new comment node class (of type TwingNodeType.COMMENT) has been added and is
extracted by the parser into the AST. Its compile method does nothing.